requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
openai
azure-search-documents
azure-core
requests
pymupdf
python-docx
python-pptx
pytesseract
Pillow
langchain-text-splitters
tiktoken
